backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.10
httpx==0.27.2
tenacity==9.0.0
pybreaker==1.1.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
